# kham_open_api_spec_2.json
{
    "openapi": "3.0.0",
    "paths": {
        "/users": {
            "post": {
                "operationId": "Users_createUser",
                "summary": "Create a new user",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The user has been successfully created."
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "Users_findAll",
                "summary": "Get all users",
                "parameters": [
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FindAllUsersResponseDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/me": {
            "get": {
                "operationId": "Users_getMe",
                "summary": "Get the current user",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "operationId": "Users_getUserById",
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "operationId": "Users_updateUser",
                "summary": "Update a user",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}/contact-info/{contactInfoId}": {
            "patch": {
                "operationId": "Users_updateContactInfo",
                "summary": "Update user contact information",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contactInfoId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateContactInfoDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}/fcm-tokens": {
            "post": {
                "operationId": "Users_addFcmToken",
                "summary": "Add a Firebase Messaging token to user profile",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddFcmTokenDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{uid}": {
            "delete": {
                "operationId": "Users_deactivateUser",
                "summary": "Deactivate a user",
                "parameters": [
                    {
                        "name": "uid",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "API",
        "description": "API DOCS",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "GeoLocation": {
                "type": "object",
                "properties": {
                    "latitude": {
                        "type": "string",
                        "example": "123"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "456"
                    }
                },
                "required": [
                    "latitude",
                    "longitude"
                ]
            },
            "CreateContactInfoDto": {
                "type": "object",
                "properties": {
                    "governorate": {
                        "type": "string",
                        "example": "Cairo"
                    },
                    "city": {
                        "type": "string",
                        "example": "Nasr City"
                    },
                    "address": {
                        "type": "string",
                        "example": "1st Area, Bld.02, Apt.01"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "example": "+20-123-456-7890"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@kham.shop"
                    },
                    "location": {
                        "$ref": "#/components/schemas/GeoLocation"
                    },
                    "default": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "governorate",
                    "city",
                    "address",
                    "phoneNumber",
                    "email"
                ]
            },
            "CreateUserDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "profileImage": {
                        "type": "string"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "business"
                        ]
                    },
                    "businessType": {
                        "type": "string",
                        "enum": [
                            "factory",
                            "supplier",
                            "restaurant"
                        ]
                    },
                    "fcmTokens": {
                        "example": [
                            "c2aK9KHmw8E:APA91bF7MY9bNnvGAXgbHN58lyDxc9KnuXNXwsqUs4uV4GyeF06HM1hMm-etu63S_4C-GnEtHAxJPJJC4H__VcIk90A69qQz65toFejxyncceg0_j5xwoFWvPQ5pzKo69rUnuCl1GSSv"
                        ],
                        "type": "array",
                        "items": {
                            "type": "array"
                        }
                    },
                    "authId": {
                        "type": "string",
                        "example": "620268d0-3f51-11ee-935a-2721f441afd1"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "contactInfoData": {
                        "$ref": "#/components/schemas/CreateContactInfoDto"
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "userType",
                    "authId"
                ]
            },
            "UserContactInfoEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "62013050-3f51-11ee-935a-2721f441afd1"
                    },
                    "userId": {
                        "type": "string",
                        "example": "62013051-3f51-11ee-935a-2721f441afd1"
                    },
                    "governorate": {
                        "type": "string",
                        "example": "Cairo"
                    },
                    "city": {
                        "type": "string",
                        "example": "Nasr City"
                    },
                    "address": {
                        "type": "string",
                        "example": "1st Area, Bld.02, Apt.01"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "example": "+20-123-456-7890"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@kham.shop"
                    },
                    "location": {
                        "$ref": "#/components/schemas/GeoLocation"
                    },
                    "default": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "userId"
                ]
            },
            "AdminAccessEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "administrator",
                            "customer_service",
                            "sales",
                            "other"
                        ]
                    },
                    "permissions": {
                        "type": "object"
                    }
                },
                "required": [
                    "id",
                    "role",
                    "permissions"
                ]
            },
            "UserEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "62013052-3f51-11ee-935a-2721f441afd1"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "profileImage": {
                        "type": "string"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "business"
                        ]
                    },
                    "businessType": {
                        "type": "string",
                        "enum": [
                            "factory",
                            "supplier",
                            "restaurant"
                        ]
                    },
                    "fcmTokens": {
                        "example": [
                            "c2aK9KHmw8E:APA91bF7MY9bNnvGAXgbHN58lyDxc9KnuXNXwsqUs4uV4GyeF06HM1hMm-etu63S_4C-GnEtHAxJPJJC4H__VcIk90A69qQz65toFejxyncceg0_j5xwoFWvPQ5pzKo69rUnuCl1GSSv"
                        ],
                        "type": "array",
                        "items": {
                            "type": "array"
                        }
                    },
                    "authId": {
                        "type": "string",
                        "example": "62013053-3f51-11ee-935a-2721f441afd1"
                    },
                    "contactInfo": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserContactInfoEntity"
                        }
                    },
                    "adminAccessId": {
                        "type": "string",
                        "example": "62013054-3f51-11ee-935a-2721f441afd1"
                    },
                    "adminAccess": {
                        "$ref": "#/components/schemas/AdminAccessEntity"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "userType",
                    "authId",
                    "isActive"
                ]
            },
            "FindAllUsersResponseDto": {
                "type": "object",
                "properties": {
                    "limit": {
                        "type": "number",
                        "readOnly": true
                    },
                    "page": {
                        "type": "number",
                        "readOnly": true
                    },
                    "data": {
                        "readOnly": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/UserEntity"
                            }
                        ]
                    }
                },
                "required": [
                    "limit",
                    "page",
                    "data"
                ]
            },
            "UserType": {
                "type": "string",
                "enum": [
                    "individual",
                    "business"
                ]
            },
            "BusinessType": {
                "type": "string",
                "enum": [
                    "factory",
                    "supplier",
                    "restaurant"
                ]
            },
            "UpdateUserDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "profileImage": {
                        "type": "string"
                    },
                    "userType": {
                        "$ref": "#/components/schemas/UserType"
                    },
                    "businessType": {
                        "$ref": "#/components/schemas/BusinessType"
                    },
                    "fcmTokens": {
                        "example": [
                            "c2aK9KHmw8E:APA91bF7MY9bNnvGAXgbHN58lyDxc9KnuXNXwsqUs4uV4GyeF06HM1hMm-etu63S_4C-GnEtHAxJPJJC4H__VcIk90A69qQz65toFejxyncceg0_j5xwoFWvPQ5pzKo69rUnuCl1GSSv"
                        ],
                        "type": "array",
                        "items": {
                            "type": "array"
                        }
                    }
                }
            },
            "AddFcmTokenDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "c2aK9KHmw8E:APA91bF7MY9bNnvGAXgbHN58lyDxc9KnuXNXwsqUs4uV4GyeF06HM1hMm-etu63S_4C-GnEtHAxJPJJC4H__VcIk90A69qQz65toFejxyncceg0_j5xwoFWvPQ5pzKo69rUnuCl1GSSv"
                    }
                },
                "required": [
                    "token"
                ]
            },
...
        }
    }
}